Move Filter and Form to new file, update all files
